Organized everything (hopefully)
Josh if you are reading this I think I have everything sorted how you sorted your content but I cant get mine to work in chrome unless all of the code is in one file. Each individual file updates on its own in chrome. I'm working on figuring it out for myself and will fix it if I get it but I'm just warming you